do not reindex as a side-effect of searching
This is bad for several reasons.

First, it's unexpected.  I'm trying to search, which should return
quickly.  Why are we reindexing, which can take a while?

Second, the implementation is incorrect.  It attempts to reindex using
the delimiter etc. passed to the Searcher, but there is no guarantee
that those are correct.  The correct values are inside the index file
which we just failed to load, so they are unavailable.

Third, it masks the cause of the problem, obfuscating it.  For example,
if the index fails to load due to ErrIndexPathMismatch, and we try to
reindex, we will fail with ErrUnknownDelimiter (unless we are lucky and
we somehow guess correctly).  For a user, ErrIndexPathMismatch is easy
enough to fix, whereas ErrUnknownDelimiter sends them the wrong way.